==> 01_Positives_5Negatives_Ages_Map.ipynb <==
import csv
import numpy as np

filename = 'data/00_NA_Clennett_deposits.csv'

lats, lons, minage = [], [], []

with open(filename) as f:
    reader = csv.reader(f)
    next(reader)
    for row in reader:
        lats.append(float(row[1]))
        lons.append(float(row[2]))
        minage.append(int(row[3]))
        
occurrences = []
non_dep_ages = []
        
occurrences = np.bincount(minage)

for n in range(1, len(occurrences)):
    if occurrences[n] == 0:
        non_dep_ages.append(n)

#print(non_dep_ages)
random_ages_nondep = np.array(non_dep_ages)
#print(random_ages_nondep)
#print(len(non_dep_ages))
#print((occurrences))

#!/usr/bin/env python3

#This is an example script of creating input file for conregistration.py.
#The input file is a text file and contains comma-separated values. 
#Each row has five fields -- index, longitude, latitude, time and plate id.

#The script contains hardcoded file names. 
#You should only use this script as an example and modify the example to prepare input data suitable to your research. 

#import os, sys
import numpy as np
#import shapefile
from parameters import parameters as param

start_time = param["time"]["start"]
end_time = param["time"]["end"]
time_step =  param["time"]["step"]

andes_points_len = len(minage)
random_ages_nondep = np.array(non_dep_ages)


positives=[] #Need to insert a first row with columns description-name. This is an issue later on
# andes deposit points with the real age
for i in range(andes_points_len):
    positives.append([i+1, lats[i], lons[i], minage[i]])


negatives_1 = []
for i in range(andes_points_len): 
    negatives_1.append([i+1, lats[i], lons[i], np.random.choice(random_ages_nondep)])
print('ready for negatives_1')

negatives_2 = []
for i in range(andes_points_len): 
    negatives_2.append([i+1, lats[i], lons[i], np.random.choice(random_ages_nondep)])
print('ready for negatives_2')
    
negatives_3 = []
for i in range(andes_points_len): 
    negatives_3.append([i+1, lats[i], lons[i], np.random.choice(random_ages_nondep)])
print('ready for negatives_3')
    
negatives_4 = []
for i in range(andes_points_len): 
    negatives_4.append([i+1, lats[i], lons[i], np.random.choice(random_ages_nondep)])
print('ready for negatives_4')
    
negatives_5 = []
for i in range(andes_points_len): 
    negatives_5.append([i+1, lats[i], lons[i], np.random.choice(random_ages_nondep)])
print('ready for negatives_5')


with open('data/01_NA_Clennett_Positives.csv',"w+") as f:
    for row in positives:
        #print(row)
        if row:
            ##f.write('{0:d}, {1:.2f}, {2:.2f}, {3:d}, {4:d}'.format(row[0],row[1],row[2],row[3],row[4]))
            f.write('{0:d}, {1:.2f}, {2:.2f}, {3:d}'.format(row[0],row[1],row[2],row[3]))
        f.write('\n')

with open('data/01_NA_Clennett_Negatives_1.csv',"w+") as f:
    for row in negatives_1:
        #print(row)
        if row:
            ##f.write('{0:d}, {1:.2f}, {2:.2f}, {3:d}, {4:d}'.format(row[0],row[1],row[2],row[3],row[4]))
            f.write('{0:d}, {1:.2f}, {2:.2f}, {3:d}'.format(row[0],row[1],row[2],row[3]))
        f.write('\n')

with open('data/01_NA_Clennett_Negatives_2.csv',"w+") as f:
    for row in negatives_2:
        #print(row)
        if row:
            ##f.write('{0:d}, {1:.2f}, {2:.2f}, {3:d}, {4:d}'.format(row[0],row[1],row[2],row[3],row[4]))
            f.write('{0:d}, {1:.2f}, {2:.2f}, {3:d}'.format(row[0],row[1],row[2],row[3]))
        f.write('\n')
        
with open('data/01_NA_Clennett_Negatives_3.csv',"w+") as f:
    for row in negatives_3:
        #print(row)
        if row:
            ##f.write('{0:d}, {1:.2f}, {2:.2f}, {3:d}, {4:d}'.format(row[0],row[1],row[2],row[3],row[4]))
            f.write('{0:d}, {1:.2f}, {2:.2f}, {3:d}'.format(row[0],row[1],row[2],row[3]))
        f.write('\n')
        
with open('data/01_NA_Clennett_Negatives_4.csv',"w+") as f:
    for row in negatives_4:
        #print(row)
        if row:
            ##f.write('{0:d}, {1:.2f}, {2:.2f}, {3:d}, {4:d}'.format(row[0],row[1],row[2],row[3],row[4]))
            f.write('{0:d}, {1:.2f}, {2:.2f}, {3:d}'.format(row[0],row[1],row[2],row[3]))
        f.write('\n')
        
with open('data/01_NA_Clennett_Negatives_5.csv',"w+") as f:
    for row in negatives_5:
        #print(row)
        if row:
            ##f.write('{0:d}, {1:.2f}, {2:.2f}, {3:d}, {4:d}'.format(row[0],row[1],row[2],row[3],row[4]))
            f.write('{0:d}, {1:.2f}, {2:.2f}, {3:d}'.format(row[0],row[1],row[2],row[3]))
        f.write('\n')

        
print('The data have been written successfully!')
