==> 02_Assigning_Plate_ID_NA_Deposits-5Negatives.ipynb <==
import pygplates
import pandas as pd
import numpy as np
from parameters import parameters
import Utils
import math
import csv


positives = pd.read_csv('data/01_NA_Clennett_Positives.csv')
negatives_1 = pd.read_csv('data/01_NA_Clennett_Negatives_1.csv')
negatives_2 = pd.read_csv('data/01_NA_Clennett_Negatives_2.csv')
negatives_3 = pd.read_csv('data/01_NA_Clennett_Negatives_3.csv')
negatives_4 = pd.read_csv('data/01_NA_Clennett_Negatives_4.csv')
negatives_5 = pd.read_csv('data/01_NA_Clennett_Negatives_5.csv')

point_features_positives = []
for index,row in positives.iterrows():
    point = pygplates.PointOnSphere(float(row[1]),float(row[2]))
    point_feature = pygplates.Feature()
    point_feature.set_geometry(point)
    point_feature.set_geometry_import_time(int(row[3]))
    point_features_positives.append(point_feature)

point_features_negatives_1 = []
for index,row in negatives_1.iterrows():
    point = pygplates.PointOnSphere(float(row[1]),float(row[2]))
    point_feature = pygplates.Feature()
    point_feature.set_geometry(point)
    point_feature.set_geometry_import_time(int(row[3]))
    point_features_negatives_1.append(point_feature)

point_features_negatives_2 = []
for index,row in negatives_2.iterrows():
    point = pygplates.PointOnSphere(float(row[1]),float(row[2]))
    point_feature = pygplates.Feature()
    point_feature.set_geometry(point)
    point_feature.set_geometry_import_time(int(row[3]))
    point_features_negatives_2.append(point_feature)


point_features_negatives_3 = []
for index,row in negatives_3.iterrows():
    point = pygplates.PointOnSphere(float(row[1]),float(row[2]))
    point_feature = pygplates.Feature()
    point_feature.set_geometry(point)
    point_feature.set_geometry_import_time(int(row[3]))
    point_features_negatives_3.append(point_feature)

point_features_negatives_4 = []
for index,row in negatives_4.iterrows():
    point = pygplates.PointOnSphere(float(row[1]),float(row[2]))
    point_feature = pygplates.Feature()
    point_feature.set_geometry(point)
    point_feature.set_geometry_import_time(int(row[3]))
    point_features_negatives_4.append(point_feature)

point_features_negatives_5 = []
for index,row in negatives_5.iterrows():
    point = pygplates.PointOnSphere(float(row[1]),float(row[2]))
    point_feature = pygplates.Feature()
    point_feature.set_geometry(point)
    point_feature.set_geometry_import_time(int(row[3]))
    point_features_negatives_5.append(point_feature)
    
static_polygons = pygplates.FeatureCollection(parameters['static_polygons_file'])
rotation_model = pygplates.RotationModel(Utils.get_files(parameters['rotation_files']))

partitioned_positives = pygplates.partition_into_plates(static_polygons, rotation_model, point_features_positives)
partitioned_negatives_1 = pygplates.partition_into_plates(static_polygons, rotation_model, point_features_negatives_1)
partitioned_negatives_2 = pygplates.partition_into_plates(static_polygons, rotation_model, point_features_negatives_2)
partitioned_negatives_3 = pygplates.partition_into_plates(static_polygons, rotation_model, point_features_negatives_3)
partitioned_negatives_4 = pygplates.partition_into_plates(static_polygons, rotation_model, point_features_negatives_4)
partitioned_negatives_5 = pygplates.partition_into_plates(static_polygons, rotation_model, point_features_negatives_5)

features_positives = pygplates.FeatureCollection(partitioned_positives)
features_negatives_1 = pygplates.FeatureCollection(partitioned_negatives_1)
features_negatives_2 = pygplates.FeatureCollection(partitioned_negatives_2)
features_negatives_3 = pygplates.FeatureCollection(partitioned_negatives_3)
features_negatives_4 = pygplates.FeatureCollection(partitioned_negatives_4)
features_negatives_5 = pygplates.FeatureCollection(partitioned_negatives_5)

one=[]
two=[]
id=1
for feature in features_positives:
    point = feature.get_geometry()
    plate_id = feature.get_reconstruction_plate_id()
    dep_age = feature.get_geometry_import_time() 
    (latitud, longitud) = point.to_lat_lon()
    
    two=[id, longitud, latitud, dep_age, plate_id]
    one.append(two)
    id=id+1
    
with open('data/02_NA_Clennett_Positives_PlateID.csv',"w+") as f:
    for row in one:
        #print(row)
        if row:
            ##f.write('{0:d}, {1:.2f}, {2:.2f}, {3:d}, {4:d}'.format(row[0],row[1],row[2],row[3],row[4]))
            f.write('{0:}, {1:.2f}, {2:.2f}, {3:}, {4:}'.format(row[0],row[1],row[2],row[3],row[4]))
        f.write('\n')


one=[]
two=[]
id=1
for feature in features_negatives_1:
    point = feature.get_geometry()
    plate_id = feature.get_reconstruction_plate_id()
    dep_age = feature.get_geometry_import_time() 
    (latitud, longitud) = point.to_lat_lon()
    
    two=[id, longitud, latitud, dep_age, plate_id]
    one.append(two)
    id=id+1
    
with open('data/02_NA_Clennett_Negatives_1_PlateID.csv',"w+") as f:
    for row in one:
        if row:
            f.write('{0:}, {1:.2f}, {2:.2f}, {3:}, {4:}'.format(row[0],row[1],row[2],row[3],row[4]))
        f.write('\n')
        
one=[]
two=[]
id=1
for feature in features_negatives_2:
    point = feature.get_geometry()
    plate_id = feature.get_reconstruction_plate_id()
    dep_age = feature.get_geometry_import_time() 
    (latitud, longitud) = point.to_lat_lon()
    
    two=[id, longitud, latitud, dep_age, plate_id]
    one.append(two)
    id=id+1
    
with open('data/02_NA_Clennett_Negatives_2_PlateID.csv',"w+") as f:
    for row in one:
        if row:
            f.write('{0:}, {1:.2f}, {2:.2f}, {3:}, {4:}'.format(row[0],row[1],row[2],row[3],row[4]))
        f.write('\n')
        
one=[]
two=[]
id=1
for feature in features_negatives_3:
    point = feature.get_geometry()
    plate_id = feature.get_reconstruction_plate_id()
    dep_age = feature.get_geometry_import_time() 
    (latitud, longitud) = point.to_lat_lon()
    
    two=[id, longitud, latitud, dep_age, plate_id]
    one.append(two)
    id=id+1
    
with open('data/02_NA_Clennett_Negatives_3_PlateID.csv',"w+") as f:
    for row in one:
        if row:
            f.write('{0:}, {1:.2f}, {2:.2f}, {3:}, {4:}'.format(row[0],row[1],row[2],row[3],row[4]))
        f.write('\n')
        

one=[]
two=[]
id=1
for feature in features_negatives_4:
    point = feature.get_geometry()
    plate_id = feature.get_reconstruction_plate_id()
    dep_age = feature.get_geometry_import_time() 
    (latitud, longitud) = point.to_lat_lon()
    
    two=[id, longitud, latitud, dep_age, plate_id]
    one.append(two)
    id=id+1
    
with open('data/02_NA_Clennett_Negatives_4_PlateID.csv',"w+") as f:
    for row in one:
        if row:
            f.write('{0:}, {1:.2f}, {2:.2f}, {3:}, {4:}'.format(row[0],row[1],row[2],row[3],row[4]))
        f.write('\n')
        

one=[]
two=[]
id=1
for feature in features_negatives_5:
    point = feature.get_geometry()
    plate_id = feature.get_reconstruction_plate_id()
    dep_age = feature.get_geometry_import_time() 
    (latitud, longitud) = point.to_lat_lon()
    
    two=[id, longitud, latitud, dep_age, plate_id]
    one.append(two)
    id=id+1
    
with open('data/02_NA_Clennett_Negatives_5_PlateID.csv',"w+") as f:
    for row in one:
        if row:
            f.write('{0:}, {1:.2f}, {2:.2f}, {3:}, {4:}'.format(row[0],row[1],row[2],row[3],row[4]))
        f.write('\n')
        
print('done')
