==> 4_Data_Wrangling-5Negatives.ipynb <==
### Step 4: Data Wrangling

The raw output data from coregistration step cannot be passed into machine learning model directly. They don't have labels and have too many features. 

The label is the thing we will try to predict with the machine learning models. In our case, the label is a True or False assertion. Each input data row is either a mineral deposit or not a mineral deposit. Given a data row, the machine learning model will tell us if the location is a mineral deposit or not. This is the ultimate goal we are trying to achieve in this machine learning workflow.

The feature is a column of the input data. There are many features in the coregistration output data. It is not wise to use too many features in the machine learning analysis because of the [Curse of dimensionality](https://en.wikipedia.org/wiki/Curse_of_dimensionality). More features mean more dimensions. If you have 20 features/columns, the machine learning model has to do the analysis in a 20-dimensional space. Just imagine if you were thrown into a 20-dimensional space, I guess it would not be a happy experience to find a way out. Yes, some machine learning models are smart enough and can reduce the number of dimensions. But at the current point of human history, humans are still a little bit smarter than computers. So, let's help computers by reducing the number of features.

We will create a csv file, in which the last colomn is the label(0 or 1) and the other columns are features. 

The feature selection is highly related to the specific research. We need to identify the features which are most important to the formation of mineral deposit. For example, some researchers might think the distance along the trench is important. Others might think the sea floor age is of great significance. Come out with your own hypothesis, wrangle the data accordingly and then send the data into a machine learning model to be evaluated. Repeat this process until we find the most important features. This process is similar to a psychic finding a perfect crystal ball to start a fortune telling business. The only difference is that the psychic is doing magic, but we are doing science.

The following code cell will select features from coregistration output and create a csv file for the machine learning analysis in Step 5.

import numpy as np
import pandas as pd
from parameters_n1 import parameters 
import Utils_c1 as Utils

#load data 
coreg_input_data = pd.read_csv('test-case-clennett/coreg_input/02_NA_Clennett_Positives_PlateID.csv')
coreg_output_data = pd.read_csv('test-case-clennett/coreg_output/02_NA_Clennett_Positives_PlateID.csv')
print('The shape of coregistration input data is: ', coreg_input_data.shape)
print('The shape of coregistration output data is: ', coreg_output_data.shape)

if coreg_input_data.shape[0] == coreg_output_data.shape[0]:
    print('Good! The input and output data has the same length ', coreg_output_data.shape[0])

print()
print('the coregistration input data')
display(coreg_input_data)
print('the coregistration output data')
display(coreg_output_data)
print('the columns in coregistration output data are: ')
Utils.print_columns()

##### After having a look at the data, let's start selecting features and add labels.

import numpy as np
import pandas as pd
from parameters_n1 import parameters 
import Utils_c1 as Utils

import os

coreg_out_dir = Utils.get_coreg_output_dir()
positive_data = pd.read_csv(coreg_out_dir + '/02_NA_Clennett_Positives_PlateID.csv')
negative_data = pd.read_csv(coreg_out_dir + '/02_NA_Clennett_Negatives_1_PlateID.csv')
candidates_data = pd.read_csv(coreg_out_dir + '/deposit_candidates.csv')

print(positive_data.columns)

feature_names = parameters['feature_names']

positive_features = positive_data[feature_names].dropna()
negative_features = negative_data[feature_names].dropna()
candidates_features = candidates_data[feature_names].dropna()

positive_features['label']=True
negative_features['label']=False

#save the data
positive_features.to_csv(Utils.get_ml_input_dir() + 'positive.csv', index=False)
negative_features.to_csv(Utils.get_ml_input_dir() + 'negative_c1.csv', index=False)
candidates_features.to_csv(Utils.get_ml_input_dir() + 'candidates.csv', index=False)

positive_data.iloc[positive_features.index].to_csv(Utils.get_ml_input_dir() + 'positive_all_columns.csv', index=False)
negative_data.iloc[negative_features.index].to_csv(Utils.get_ml_input_dir() + 'negative_c1_all_columns.csv', index=False)
candidates_data.iloc[candidates_features.index].to_csv(Utils.get_ml_input_dir() + '/candidates_all_columns.csv', index=False)

import glob
files = glob.glob(Utils.get_ml_input_dir() + '*')
print('\ngenerated files:')
for f in files:
    print(f)


import numpy as np
import pandas as pd
from parameters_n2 import parameters 
import Utils_c2 as Utils

import os

coreg_out_dir = Utils.get_coreg_output_dir()
positive_data = pd.read_csv(coreg_out_dir + '/02_NA_Clennett_Positives_PlateID.csv')
negative_data = pd.read_csv(coreg_out_dir + '/02_NA_Clennett_Negatives_2_PlateID.csv')
candidates_data = pd.read_csv(coreg_out_dir + '/deposit_candidates.csv')

print(positive_data.columns)

feature_names = parameters['feature_names']

positive_features = positive_data[feature_names].dropna()
negative_features = negative_data[feature_names].dropna()
candidates_features = candidates_data[feature_names].dropna()

positive_features['label']=True
negative_features['label']=False

#save the data
positive_features.to_csv(Utils.get_ml_input_dir() + 'positive.csv', index=False)
negative_features.to_csv(Utils.get_ml_input_dir() + 'negative_c2.csv', index=False)
candidates_features.to_csv(Utils.get_ml_input_dir() + 'candidates.csv', index=False)

positive_data.iloc[positive_features.index].to_csv(Utils.get_ml_input_dir() + 'positive_all_columns.csv', index=False)
negative_data.iloc[negative_features.index].to_csv(Utils.get_ml_input_dir() + 'negative_c2_all_columns.csv', index=False)
candidates_data.iloc[candidates_features.index].to_csv(Utils.get_ml_input_dir() + '/candidates_all_columns.csv', index=False)

import glob
files = glob.glob(Utils.get_ml_input_dir() + '*')
print('\ngenerated files:')
for f in files:
    print(f)

import numpy as np
import pandas as pd
from parameters_n3 import parameters 
import Utils_c3 as Utils

import os

coreg_out_dir = Utils.get_coreg_output_dir()
positive_data = pd.read_csv(coreg_out_dir + '/02_NA_Clennett_Positives_PlateID.csv')
negative_data = pd.read_csv(coreg_out_dir + '/02_NA_Clennett_Negatives_3_PlateID.csv')
candidates_data = pd.read_csv(coreg_out_dir + '/deposit_candidates.csv')

print(positive_data.columns)

feature_names = parameters['feature_names']

positive_features = positive_data[feature_names].dropna()
negative_features = negative_data[feature_names].dropna()
candidates_features = candidates_data[feature_names].dropna()

positive_features['label']=True
negative_features['label']=False

#save the data
positive_features.to_csv(Utils.get_ml_input_dir() + 'positive.csv', index=False)
negative_features.to_csv(Utils.get_ml_input_dir() + 'negative_c3.csv', index=False)
candidates_features.to_csv(Utils.get_ml_input_dir() + 'candidates.csv', index=False)

positive_data.iloc[positive_features.index].to_csv(Utils.get_ml_input_dir() + 'positive_all_columns.csv', index=False)
negative_data.iloc[negative_features.index].to_csv(Utils.get_ml_input_dir() + 'negative_c3_all_columns.csv', index=False)
candidates_data.iloc[candidates_features.index].to_csv(Utils.get_ml_input_dir() + '/candidates_all_columns.csv', index=False)

import glob
files = glob.glob(Utils.get_ml_input_dir() + '*')
print('\ngenerated files:')
for f in files:
    print(f)

import numpy as np
import pandas as pd
from parameters_n4 import parameters 
import Utils_c4 as Utils

import os

coreg_out_dir = Utils.get_coreg_output_dir()
positive_data = pd.read_csv(coreg_out_dir + '/02_NA_Clennett_Positives_PlateID.csv')
negative_data = pd.read_csv(coreg_out_dir + '/02_NA_Clennett_Negatives_4_PlateID.csv')
candidates_data = pd.read_csv(coreg_out_dir + '/deposit_candidates.csv')

print(positive_data.columns)

feature_names = parameters['feature_names']

positive_features = positive_data[feature_names].dropna()
negative_features = negative_data[feature_names].dropna()
candidates_features = candidates_data[feature_names].dropna()

positive_features['label']=True
negative_features['label']=False

#save the data
positive_features.to_csv(Utils.get_ml_input_dir() + 'positive.csv', index=False)
negative_features.to_csv(Utils.get_ml_input_dir() + 'negative_c4.csv', index=False)
candidates_features.to_csv(Utils.get_ml_input_dir() + 'candidates.csv', index=False)

positive_data.iloc[positive_features.index].to_csv(Utils.get_ml_input_dir() + 'positive_all_columns.csv', index=False)
negative_data.iloc[negative_features.index].to_csv(Utils.get_ml_input_dir() + 'negative_c4_all_columns.csv', index=False)
candidates_data.iloc[candidates_features.index].to_csv(Utils.get_ml_input_dir() + '/candidates_all_columns.csv', index=False)

import glob
files = glob.glob(Utils.get_ml_input_dir() + '*')
print('\ngenerated files:')
for f in files:
    print(f)

import numpy as np
import pandas as pd
from parameters_n5 import parameters 
import Utils_c5 as Utils

import os

coreg_out_dir = Utils.get_coreg_output_dir()
positive_data = pd.read_csv(coreg_out_dir + '/02_NA_Clennett_Positives_PlateID.csv')
negative_data = pd.read_csv(coreg_out_dir + '/02_NA_Clennett_Negatives_5_PlateID.csv')
candidates_data = pd.read_csv(coreg_out_dir + '/deposit_candidates.csv')

print(positive_data.columns)

feature_names = parameters['feature_names']

positive_features = positive_data[feature_names].dropna()
negative_features = negative_data[feature_names].dropna()
candidates_features = candidates_data[feature_names].dropna()

positive_features['label']=True
negative_features['label']=False

#save the data
positive_features.to_csv(Utils.get_ml_input_dir() + 'positive.csv', index=False)
negative_features.to_csv(Utils.get_ml_input_dir() + 'negative_c5.csv', index=False)
candidates_features.to_csv(Utils.get_ml_input_dir() + 'candidates.csv', index=False)

positive_data.iloc[positive_features.index].to_csv(Utils.get_ml_input_dir() + 'positive_all_columns.csv', index=False)
negative_data.iloc[negative_features.index].to_csv(Utils.get_ml_input_dir() + 'negative_c5_all_columns.csv', index=False)
candidates_data.iloc[candidates_features.index].to_csv(Utils.get_ml_input_dir() + '/candidates_all_columns.csv', index=False)

import glob
files = glob.glob(Utils.get_ml_input_dir() + '*')
print('\ngenerated files:')
for f in files:
    print(f)
